+++ parsley/src/ghc/Parsley/Combinator.hs
@@ -1,5 +1,3 @@
-{-# OPTIONS_GHC -Wno-incomplete-patterns #-}
-{-# LANGUAGE PatternSynonyms, ViewPatterns #-}
 {-|
 Module      : Parsley.Combinator
 Description : The parsing combinators
@@ -7,15 +5,12 @@ License     : BSD-3-Clause
 Maintainer  : Jamie Willis
 Stability   : stable
 
-This module contains the classic parser combinator operations specific to parsers themselves.
-This means any combinators that deal with input consumption at a primitive level.
+This module contains the classic parser combinator operations specific to parsers themselves, excluding
+those that consume input.
 
 @since 0.1.0.0
 -}
 module Parsley.Combinator (
-    satisfy, char, item,
-    string, token,
-    oneOf, noneOf,
     eof, more,
     someTill,
     try,
@@ -23,13 +18,11 @@ module Parsley.Combinator (
     line, col, pos
   ) where
 
-import Prelude hiding           (traverse, (*>))
-import Data.List                (sort)
+import Prelude hiding           ((*>))
 import Parsley.Alternative      (manyTill)
-import Parsley.Applicative      (($>), void, traverse, (<:>), (*>), (<~>))
-import Parsley.Defunctionalized (Defunc(LIFTED, RANGES))
+import Parsley.Applicative      (void, (<:>), (*>), (<~>))
 import Parsley.Internal         (Parser)
-import Parsley.ParserOps        (satisfy)
+import Parsley.Char             (item)
 
 import qualified Parsley.Internal as Internal (try, lookAhead, notFollowedBy, line, col)
 
@@ -70,55 +63,6 @@ try :: Parser a -> Parser a
 try = Internal.try
 
 {-|
-This combinator will attempt match a given string. If the parser fails midway through, this
-combinator will fail having consumed input. On success, the string itself is returned and input
-will be consumed.
-
-@since 0.1.0.0
--}
-string :: String -> Parser String
-string = traverse char
-
-{-|
-This combinator will attempt to match any one of the provided list of characters. If one of those
-characters is found, it will be returned and the input consumed. If not, the combinator will fail
-having consumed no input.
-
-@since 0.1.0.0
--}
-oneOf :: [Char] -> Parser Char
-oneOf = satisfy . RANGES True . ranges
-
-{-|
-This combinator will attempt to not match any one of the provided list of characters. If one of those
-characters is found, the combinator will fail having consumed no input. If not, it will return
-the character that was not an element of the provided list.
-
-@since 0.1.0.0
--}
-noneOf :: [Char] -> Parser Char
-noneOf = satisfy . RANGES False . ranges
-
-ranges :: [Char] -> [(Char, Char)]
-ranges [] = []
-ranges (sort -> c:cs) = go c (fromEnum c) cs
-  where
-    go :: Char -> Int -> [Char] -> [(Char, Char)]
-    go lower prev [] = [(lower, toEnum prev)]
-    go lower prev (c:cs)
-      | i <- fromEnum c, i == prev + 1 = go lower i cs
-      | otherwise = (lower, toEnum prev) : go c (fromEnum c) cs
-
-{-|
-Like `string`, excepts parses the given string atomically using `try`. Never consumes input on
-failure.
-
-@since 0.1.0.0
--}
-token :: String -> Parser String
-token = try . string
-
-{-|
 This parser succeeds only if there is no input left to consume, and fails without consuming input
 otherwise.
 
@@ -134,26 +78,6 @@ This parser succeeds if there is still input left to consume, and fails otherwis
 -}
 more :: Parser ()
 more = lookAhead (void item)
-
--- Parsing Primitives
-{-|
-This combinator will attempt to match a given character. If that character is the next input token,
-the parser succeeds and the character is returned. Otherwise, the combinator will fail having not
-consumed any input.
-
-@since 0.1.0.0
--}
-char :: Char -> Parser Char
-char c = satisfy (RANGES True [(c, c)]) $> LIFTED c
-
-{-|
-Reads any single character. This combinator will only fail if there is no more input remaining.
-The parsed character is returned.
-
-@since 0.1.0.0
--}
-item :: Parser Char
-item = satisfy (RANGES False [])
 
 -- Composite Combinators
 {-|